fixes #2061 Move IPC parameters from uint64 to int

+++ src/platform/posix/posix_peerid.h
@@ -1,5 +1,5 @@
 //
-// Copyright 2023 Staysail Systems, Inc. <info@staysail.tech>
+// Copyright 2025 Staysail Systems, Inc. <info@staysail.tech>
 //
 // This software is supplied under the terms of the MIT License, a
 // copy of which should be located in the distribution where this
@@ -19,7 +19,6 @@
 #include "core/nng_impl.h"
 #include <sys/types.h>
 
-int nni_posix_peerid(
-    int fd, uint64_t *euid, uint64_t *egid, uint64_t *prid, uint64_t *znid);
+int nni_posix_peerid(int fd, int *euid, int *egid, int *prid, int *znid);
 
 #endif // PLATFORM_POSIX_PEERID_H
